requirements.txt
streamlit==1.31.0
pandas==2.1.1
matplotlib==3.8.0
python-dotenv==1.0.1
requests==2.31.0
boto3>=1.28.57
superlinked==18.5.0
umap-learn==0.5.7
scikit-learn==1.6.0
altair>=5.0.0,<6.0.0
superlinked-server==0.7.4
